membercard shows study and period participated in fondet
