+++ src/app/admin/layout.tsx
@@ -1,0 +1,77 @@
+"use client";
+
+import Link from "next/link";
+import { usePathname, useRouter } from "next/navigation";
+
+const navItems = [
+  { href: "/admin", label: "Oversikt" },
+  { href: "/admin/medlemmer", label: "Medlemmer" },
+  { href: "/admin/rapporter", label: "Rapporter" },
+  { href: "/admin/soknader", label: "Søknader" },
+];
+
+export default function AdminLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  const pathname = usePathname();
+  const router = useRouter();
+
+  if (pathname === "/admin/login") {
+    return <>{children}</>;
+  }
+
+  async function handleLogout() {
+    await fetch("/api/auth/logout", { method: "POST" });
+    router.push("/");
+  }
+
+  return (
+    <div className="min-h-screen bg-gray-950 pt-24 overflow-x-hidden">
+      <div className="w-full max-w-7xl mx-auto px-8">
+        <div className="flex justify-between items-center mb-4 pt-6">
+          <h1 className="text-3xl font-bold text-white">Admin Dashboard</h1>
+          <div className="flex gap-3">
+            <a
+              href="/"
+              className="px-4 py-2 bg-blue-600 text-white rounded hover:bg-blue-700 transition"
+            >
+              Til forsiden
+            </a>
+            <button
+              onClick={handleLogout}
+              className="px-4 py-2 bg-red-600 text-white rounded hover:bg-red-700 transition"
+            >
+              Logg ut
+            </button>
+          </div>
+        </div>
+
+        <nav className="flex gap-1 mb-8 border-b border-gray-800">
+          {navItems.map((item) => {
+            const isActive =
+              item.href === "/admin"
+                ? pathname === "/admin"
+                : pathname.startsWith(item.href);
+            return (
+              <Link
+                key={item.href}
+                href={item.href}
+                className={`px-4 py-2 text-sm font-medium rounded-t transition-colors ${
+                  isActive
+                    ? "bg-gray-800 text-white border-b-2 border-blue-500"
+                    : "text-gray-400 hover:text-white hover:bg-gray-900"
+                }`}
+              >
+                {item.label}
+              </Link>
+            );
+          })}
+        </nav>
+
+        <div className="pb-8">{children}</div>
+      </div>
+    </div>
+  );
+}
