+++ src/app/admin/soknader/page.tsx
@@ -1,0 +1,12 @@
+"use client";
+
+export default function SoknaderPage() {
+  return (
+    <div>
+      <h2 className="text-xl font-semibold text-white mb-4">Søknader</h2>
+      <p className="text-gray-400">
+        Administrering av søknader kommer snart. Søknader sendes foreløpig via e-post til fondet@tihlde.org.
+      </p>
+    </div>
+  );
+}
